Added support for reapplying additional asset type extras
- combination cores are skipped at the moment
- refactored pocket extras code into service to allow for check on core update
- corrected asset prefix typo in JSON
- removed the unzipped jotego beta key when finished
