+++ pocket_extras.json
@@ -82,7 +82,7 @@
       "has_placeholders": true,
       "github_user": "dyreschlock",
       "github_repository": "pocket-extras",
-      "github_asset_prefix": "pocket_extras_toaplan2_c"
+      "github_asset_prefix": "pocket-extras-toaplan2_c"
     },
     {
       "id": "jts16_c",
